requirements.txt
PyMuPDF
tqdm
Pillow==11.2.1
zhipuai